Rename model --> types
